Exact Expected SFS
Computes exact expected branch length spectrum for a general Lambda-coalescent, in particular  for Example 2.4 in Durrett and Schweinsberg, a coalescent model for a genealogy at a single site in a population of size 2N evolving according to the Moran model with  strong selection, i.e. there are repeated beneficial mutations that have a significant chance of sweeping to fixation in roughly log(2N)  generations on average.